moved internal helper functions downward
I had forgotten about the implications of hoisting (i.e., all function declarations are moved to the top of their scope when the interpreter reads the file). This means that helper functions can be accessed higher up in the file (unlike variable declarations within a function which go sequentially), thus allowing for cleaner util files where the main export is at the top of the messy internals down below.